refactor: Use standard setup
