+++ examples/gofer-3/test/.eslintrc
@@ -1,0 +1,1 @@
+../../../test/.eslintrc
